test(mistral4): keep HF precision adjustment in checkpoint harness
