+++ tests/ci_tests/README.md
@@ -258,14 +258,15 @@ profile and numeric override mappings remain available for measured one-model ex
 
 ### Explicit HF reference precision
 
-`ci.checkpoint_robustness.hf_reference_context` optionally supplies a model-owned context manager using the existing
-`_target_` configuration syntax. The harness passes the loaded HF model as `model`; the context surrounds source and
-export-reload forwards, including repeatability and shape diagnostics, and is closed even if a forward fails.
-Omitting it retains the native HF compute policy. Configured contexts are logged explicitly as modified references.
+`ci.checkpoint_robustness.hf_router_scores_fp32: true` enables a test-harness adjustment alongside the existing
+FP32 tensor-loading treatment. It currently supports HF Mistral4 routers and rejects models without those routers.
+The original HF forward uses FP32 softmax, normalization, and selected weights while retaining native projection
+and checkpoint-storage dtypes. Tensor preservation acts at load time; this option controls scoring at forward time.
 
-The Mistral4 recipe selects `nemo_automodel.components.models.mistral4.reference_precision.fp32_router_scores`.
-It executes the original HF router with FP32 softmax, normalization, and selected weights, while retaining native
-projection and checkpoint-storage dtypes. It affects neither AutoModel training nor other HF modules. This is a
+The harness applies the adjustment to source and export-reload forwards, including repeatability and shape
+diagnostics, and restores the original instance forwards even if a comparison fails. Omitting the option or setting
+it to false retains native HF computation. Enabling it logs an explicit modified-reference message. All adjustment
+code lives in the checkpoint test harness; it affects neither AutoModel training nor other HF modules. This is a
 precision-controlled comparison, not proof of the author's intended training precision or of full-model parity.
 Source and reload thresholds are still enforced; reference precision settings must not implicitly relax them.
