+++ tests/unit_tests/ci_tests/test_checkpoint_robustness_hf_reference_precision.py
@@ -18,11 +18,14 @@ from transformers.models.mistral4.configuration_mistral4 import Mistral4Config
 from transformers.models.mistral4.modeling_mistral4 import Mistral4TopkRouter
 
 from nemo_automodel.components.config.loader import ConfigNode
-from nemo_automodel.components.models.mistral4.reference_precision import fp32_router_scores
 from tests.functional_tests.checkpoint_robustness.test_checkpoint_robustness_llm import (
     _extract_custom_args,
     _hf_reference_context,
 )
+
+
+def _reference_config(enabled: bool = True) -> ConfigNode:
+    return ConfigNode({"ci": {"checkpoint_robustness": {"hf_router_scores_fp32": enabled}}})
 
 
 def _router(dtype=torch.bfloat16):
@@ -42,7 +45,7 @@ def test_fp32_scores_preserve_projection_and_match_weights_and_gradients(dtype):
     original_weight = router.weight
     original_values = router.weight.detach().clone()
 
-    with fp32_router_scores(router):
+    with _hf_reference_context(_reference_config(), router):
         logits, weights, indices = router(hidden)
         assert logits.dtype == dtype
         assert weights.dtype == torch.float32
@@ -87,7 +90,7 @@ def test_reference_context_preserves_existing_instance_wrapper_and_cleans_up_aft
     unaffected = _router()
     hidden = torch.ones(2, 16, dtype=torch.bfloat16)
     with pytest.raises(RuntimeError, match="sentinel"):
-        with fp32_router_scores(router):
+        with _hf_reference_context(_reference_config(), router):
             assert router(hidden)[1].dtype == torch.float32
             assert unaffected(hidden)[1].dtype == torch.bfloat16
             raise RuntimeError("sentinel")
@@ -113,27 +116,27 @@ def test_reference_context_fails_if_hf_stops_using_the_expected_softmax():
     router.forward = changed_forward
     original = router.forward
     with pytest.raises(RuntimeError, match="softmax contract changed"):
-        with fp32_router_scores(router):
+        with _hf_reference_context(_reference_config(), router):
             router(torch.ones(2, 16, dtype=torch.bfloat16))
     assert router.forward is original
 
 
 def test_reference_context_rejects_other_model_families():
     with pytest.raises(ValueError, match="no HF Mistral4 routers"):
-        with fp32_router_scores(torch.nn.Linear(16, 8)):
+        with _hf_reference_context(_reference_config(), torch.nn.Linear(16, 8)):
             pass
 
 
 def test_tracked_recipe_selects_precision_context_without_relaxing_source_gates():
     from pathlib import Path
 
-    recipe = Path(__file__).resolve().parents[4] / "examples/vlm_finetune/mistral4/mistral4_medpix.yaml"
+    recipe = Path(__file__).resolve().parents[3] / "examples/vlm_finetune/mistral4/mistral4_medpix.yaml"
     import yaml
 
     cfg = ConfigNode(yaml.safe_load(recipe.read_text()))
     custom, remaining = _extract_custom_args(["--config", str(recipe)])
-    assert "hf_reference_context" not in custom
-    assert not any("hf_reference_context" in value for value in remaining)
+    assert "hf_router_scores_fp32" not in custom
+    assert not any("hf_router_scores_fp32" in value for value in remaining)
     assert "source_load" not in custom["parity_tolerance_profile_overrides"]
     assert "parity_threshold_overrides" not in custom
     router = _router()
@@ -143,3 +146,11 @@ def test_tracked_recipe_selects_precision_context_without_relaxing_source_gates(
     assert router(hidden)[1].dtype == torch.bfloat16
     with _hf_reference_context(ConfigNode({}), router):
         assert router(hidden)[1].dtype == torch.bfloat16
+
+
+def test_reference_precision_can_be_disabled_explicitly():
+    router = _router()
+    hidden = torch.ones(2, 16, dtype=torch.bfloat16)
+    with _hf_reference_context(_reference_config(enabled=False), router):
+        assert router(hidden)[1].dtype == torch.bfloat16
+    assert "forward" not in router.__dict__
